feat: auth role 完善
1. auth role 文档配置
2. 修复auth role中的api

+++ template/adminauthtpl/tpl/core/app/auth/api/delete.tpl
@@ -22,16 +22,18 @@ func NewAuthDel(c *gin.Context, sc *ctx.ServerContext) irouter.IHandler {
 
 //	@tags		权限管理
 //	@summary	删除权限
+//	@x-group	{"key":"auth","inorder":2}
 //	@router		/api/auth/:id [delete]
-//	@param		id				path		int							true	"用户ID"
-//	@param		Authorization	header		string						true	"token"
-//	@success	200				{object}	hd.Response{data=string}	"ok"
+//	@param		Authorization	header		string	true	" "	extensions(x-name=鉴权,x-value=[TOKEN])
+//	@param		id				path		int		true	" "	extensions(x-name=用户ID,x-value=1)
+//	@success	200				{string}	string	"{data:'ok'}"
 func (this *AuthDel) Do() error {
 	var err error
 	id, err := this.GetId()
 	if err != nil {
 		return err
 	}
+
 	if err := this.Delete(id); err != nil {
 		return err
 	} else {
@@ -52,6 +54,15 @@ func (this *AuthDel) Delete(id int64) error {
 			return r.Error
 		}
 	}
+	ct := int64(0)
+	if r := db.Model(&authmodel.AuthPo{}).Where("parent_id=?", id).Count(&ct); r.Error != nil {
+		return r.Error
+	}
+
+	if ct != 0 {
+		return errors.New("该权限下包含其他权限，禁止删除！")
+	}
+
 	if r := db.Delete(po); r.Error != nil {
 		return r.Error
 	}
